+++ src/DotRecast.Recast.Toolset/Tools/RcTestNavMeshTool.cs
@@ -380,7 +380,7 @@ namespace DotRecast.Recast.Toolset.Tools
             float nx = (epos.Z - spos.Z) * 0.25f;
             float nz = -(epos.X - spos.X) * 0.25f;
 
-            var tempQueryPoly = new RcVec3f[4];
+            RcVec3f[] tempQueryPoly = new RcVec3f[4];
             tempQueryPoly[0].X = spos.X + nx * 1.2f;
             tempQueryPoly[0].Y = spos.Y + agentHeight / 2;
             tempQueryPoly[0].Z = spos.Z + nz * 1.2f;